+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.7.0
 
-- Added: `swiftlint.additionalParameters` to enable `--format` fixes [vknabel/vscode-swiftformat#43](https://github.com/vknabel/vscode-swiftformat/issues/43)
+- Added: `swiftlint.additionalParameters` to enable `--format` fixes #43
+- Added: build SwiftLint if needed #42
 
 ## 1.6.0